requirements.txt
requests
kafka-python
schedule
pyspark
google-cloud-storage
flask
web3
py-evm